requirements.txt
pycryptodome
flask
ImageHash